+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Loader: MuJoCo
+- Unit tests to check that MuJoCo loads all MJCF descriptions successfully
 
 ## [0.4.0] - 2022/09/14