+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - Cassie: update to description from MuJoCo Menagerie
+- Pinocchio loader: no root joint by default
 
 ## [0.4.0] - 2022/09/14